tracker_app/backend/requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Computer Vision
opencv-python>=4.9.0
numpy>=1.26.0
pycocotools>=2.0.7

# Detection (choose one or both)
ultralytics>=8.2.0          # YOLOv8 (recommended)
torchvision>=0.17.0         # FasterRCNN, SSD MobileNet
# torch>=2.2.0               # For YOLOv5 via torch.hub

# ── Tracking Frameworks ───────────────────────────────────────────────────────

# BOXMOT (includes ByteTrack, BoT-SORT, OC-SORT, Deep OC-SORT, StrongSORT)
boxmot>=10.0.60

# DeepSORT
deep-sort-realtime>=1.3.2

# FairMOT — install from source:
#   pip install git+https://github.com/ifzhang/FairMOT.git
# or PyPI fork:
#   pip install fairmot

# RE-ID (used by boxmot/deepsort)
# torchreid is bundled with boxmot; for standalone:
#   pip install torchreid

# ── Optional ─────────────────────────────────────────────────────────────────
# sort-python>=1.1.0         # Fallback pure SORT tracker
# lap>=0.4.0                 # Hungarian algorithm (required by some trackers)
